lending: Add overflow check in proptest (#1199)
This issue was discovered during CI:
https://github.com/solana-labs/solana-program-library/runs/1833196857

The approach is to do the same calc as `_liquidate_obligation` and make
sure that the `MockConverter` doesn't overflow `Decimal`.  This seemed
like a simpler approach than doing a complex local proptest constraint.

On top of that, it includes the proptest regression to run it on every
`cargo test`.