token 2022 test: refactor transfer hook test to use new offchain helper
Following up from the new helpers added in #6099 and #6100, there's a transfer
hook test in `program-2022-test` that's using the now-deprecated helper to build
a swap instruction for the test program.

This PR replaces that bit of code to use the new helper from
`spl_transfer_hook_interface`.

Note: The test in question is designed to test Token2022's
`invoke_transfer_checked(..)` on-chain helper, *not* the offchain helpers. So,
it's appropriate to use the offchain helper directly from the interface crate
for this step.